Update blinky_console.md
I'm not sure if I don't understand something, or if this is a coding error.  I'm still learning about Callouts, but it seems that "os_callout_init" only has three parameters according to the documentation whereas, "os_callout_func_init" has four.  Additionally, this code seems to be desiring to call the "timer_ev_cb" function whereas "os_callout_init" is only suppose to send a parameter in.  Thoughts?